+++ src/main/java/com/nisovin/shopkeepers/util/TextUtils.java
@@ -37,8 +37,10 @@ public class TextUtils {
 	// FORMATTING AND CONVERSION
 
 	public static final DecimalFormat DECIMAL_FORMAT = new DecimalFormat("0.##", new DecimalFormatSymbols(Locale.US));
+	public static final DecimalFormat DECIMAL_FORMAT_PRECISE = new DecimalFormat("0.####", new DecimalFormatSymbols(Locale.US));
 	static {
 		DECIMAL_FORMAT.setGroupingUsed(false);
+		DECIMAL_FORMAT_PRECISE.setGroupingUsed(false);
 	}
 
 	public static String getLocationString(Location location) {